api-call/requirements.txt
google-ads==13.0.0
mysql-connector-python==8.0.25
requests==2.25.1
pyyaml==5.4.1
